creation of cogs folder
